// src/enums/scrapers/extractor.ts
export enum Extractor {
    InnerText = "innerText",
    TextContent = "textContent",
    Href = "href",
    ClassList = "classList",
    ClassNames = "classNames",
    Src = "src",
    AriaLabel = "ariaLabel",
    AriaHidden = "ariaHidden",
    AriaChecked = "ariaChecked",
    AriaExpanded = "ariaExpanded",
    AriaDisabled = "ariaDisabled",
    Attrs = "attrs",
    Attributes = "attributes",
    RawAttrs = "rawAttrs",
    RawAttributes = "rawAttributes",
    RawText = "rawText",
    OuterHtml = "outerHtml",
    InnerHtml = "innerHtml",
    TagName = "tagName",
    RawTagName = "rawTagName",
    Id = "id",
    Type = "type",
    Role = "role",
    Name = "name",
    Placeholder = "placeholder",
    Title = "title",
    Value = "value",
    Checked = "checked",
    Disabled = "disabled",
    Readonly = "readonly",
    Required = "required",
    Selected = "selected",
    NodeType = "nodeType",
    LocalName = "localName",
    Autocomplete = "autocomplete",
    MaxLength = "maxLength",
    MinLength = "minLength",
    ChildElementCount = "childElementCount",
    Pattern = "pattern",
    Style = "style",
    Text = "text",
    Structure = "structure",
    StructuredText = "structuredText",
    Data = "data",
    Attribute = "attribute",
    Range = "range",
    Content = "content",
}